hw7-apache-beam/requirements.txt

wheel
apache-beam[gcp]
google-cloud-storage